chore: project structure refactoring (styling, data)
