+++ components/Footer.tsx
@@ -1,104 +1,9 @@
-import { createStyles, Text, Container, ActionIcon, Group } from '@mantine/core';
+import { Text, Container, ActionIcon, Group } from '@mantine/core';
 import { IconBrandTwitter, IconBrandGithub, IconBrandInstagram } from '@tabler/icons';
 import Link from 'next/link';
 import Image from 'next/image';
 import React from 'react';
-
-const useStyles = createStyles((theme) => ({
-    footer: {
-        marginTop: 120,
-        paddingTop: theme.spacing.xl * 2,
-        paddingBottom: theme.spacing.xl * 2,
-        backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
-        borderTop: `1px solid ${
-          theme.colorScheme === 'dark' ? theme.colors.dark[5] : theme.colors.gray[2]
-        }`,
-    },
-
-    logo: {
-        maxWidth: 200,
-
-        [theme.fn.smallerThan('sm')]: {
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-        },
-    },
-
-    description: {
-        marginTop: 5,
-
-        [theme.fn.smallerThan('sm')]: {
-            marginTop: theme.spacing.xs,
-            textAlign: 'center',
-        },
-    },
-
-    inner: {
-        display: 'flex',
-        justifyContent: 'space-between',
-
-        [theme.fn.smallerThan('sm')]: {
-            flexDirection: 'column',
-            alignItems: 'center',
-        },
-    },
-
-    groups: {
-        display: 'flex',
-        flexWrap: 'wrap',
-
-        [theme.fn.smallerThan('sm')]: {
-            display: 'none',
-        },
-    },
-
-    wrapper: {
-        width: 160,
-    },
-
-    link: {
-        display: 'block',
-        color: theme.colorScheme === 'dark' ? theme.colors.dark[1] : theme.colors.gray[6],
-        fontSize: theme.fontSizes.sm,
-        paddingTop: 3,
-        paddingBottom: 3,
-
-        '&:hover': {
-            textDecoration: 'underline',
-        },
-    },
-
-    title: {
-        fontSize: theme.fontSizes.lg,
-        fontWeight: 700,
-        fontFamily: `Greycliff CF, ${theme.fontFamily}`,
-        marginBottom: theme.spacing.xs / 2,
-        color: theme.colorScheme === 'dark' ? theme.white : theme.black,
-    },
-
-    afterFooter: {
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        marginTop: theme.spacing.xl,
-        paddingTop: theme.spacing.xl,
-        paddingBottom: theme.spacing.xl,
-        borderTop: `1px solid ${
-          theme.colorScheme === 'dark' ? theme.colors.dark[4] : theme.colors.gray[2]
-        }`,
-
-        [theme.fn.smallerThan('sm')]: {
-            flexDirection: 'column',
-        },
-    },
-
-    social: {
-        [theme.fn.smallerThan('sm')]: {
-            marginTop: theme.spacing.xs,
-        },
-    },
-}));
+import { useStyles } from '../css/style';
 
 interface FooterLinksProps {
     data: {
@@ -114,7 +19,7 @@ export function FooterLinks({ data }: FooterLinksProps) {
         const links = group.links.map((link, index) => (
           <Text<'a'>
             key={index}
-            className={classes.link}
+            className={classes.footerLink}
             component="a"
             href={link.link}
             onClick={(event) => event.preventDefault()}
@@ -142,9 +47,6 @@ export function FooterLinks({ data }: FooterLinksProps) {
                     width={75}
                     alt="CodeQuantum logo"
                   />
-                  {/*<Text size="xs" color="dimmed" className={classes.description}>*/}
-                  {/*    TODO: see if CodeQuantum has a motto for this spot*/}
-                  {/*</Text>*/}
               </div>
               <div className={classes.groups}>
                   {groups}
