+++ components/Header.tsx
@@ -1,55 +1,8 @@
 import React, { useState } from 'react';
-import { createStyles, Header, Container, Group, Modal } from '@mantine/core';
+import { Header, Container, Group, Modal } from '@mantine/core';
 import { IconMenu2 } from '@tabler/icons';
 import Image from 'next/image';
-
-const useStyles = createStyles((theme) => ({
-    header: {
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        height: '100%',
-        border: 'none',
-    },
-
-    links: {
-        [theme.fn.smallerThan('sm')]: {
-            display: 'none',
-        },
-    },
-
-    burger: {
-        [theme.fn.largerThan('sm')]: {
-            display: 'none',
-        },
-        width: '30px',
-    },
-
-    link: {
-        display: 'block',
-        lineHeight: 1,
-        padding: '8px 12px',
-        borderRadius: theme.radius.sm,
-        textDecoration: 'none',
-        color: theme.colorScheme === 'dark' ? theme.colors.dark[0] : theme.colors.gray[7],
-        fontSize: theme.fontSizes.sm,
-        fontWeight: 500,
-
-        '&:hover': {
-            backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
-        },
-    },
-
-    linkActive: {
-        '&, &:hover': {
-            backgroundColor:
-                theme.colorScheme === 'dark'
-                    ? theme.fn.rgba(theme.colors[theme.primaryColor][9], 0.25)
-                    : theme.colors[theme.primaryColor][0],
-            color: theme.colors[theme.primaryColor][theme.colorScheme === 'dark' ? 3 : 7],
-        },
-    },
-}));
+import { useStyles } from '../css/style';
 
 interface HeaderSimpleProps {
     links: { link: string; label: string }[];
@@ -69,6 +22,19 @@ export function HeaderSimple({ links }: HeaderSimpleProps) {
                 event.preventDefault();
                 setActive(link.link);
             }}
+        >
+            {link.label}
+        </a>
+    ));
+
+    const modalItems = links.map((link) => (
+        <a
+          key={link.label}
+          href={link.link}
+          className={cx(classes.modalLink, { [classes.linkActive]: active === link.link })}
+          onClick={() => {
+              setActive(link.link);
+          }}
         >
             {link.label}
         </a>
@@ -95,13 +61,14 @@ export function HeaderSimple({ links }: HeaderSimpleProps) {
                   size="sm"
                 />
             </Container>
+
             <Modal
               opened={opened}
               onClose={() => toggleOpened(!opened)}
               title="CodeQuantum 2022"
               fullScreen
             >
-                {items}
+                {modalItems}
             </Modal>
         </Header>
     );
