+++ components/Sponsors.tsx
@@ -4,62 +4,16 @@ import {
   Center,
   Container,
   SimpleGrid,
-  createStyles,
 } from '@mantine/core';
-
-export const sponsorList = [
-  {
-    imagePath: '/images/sponsors/google-cloud.png',
-    altText: 'Google Cloud',
-  },
-  {
-    imagePath: '/images/sponsors/H-E-B.png',
-    altText: 'H-E-B',
-  },
-  {
-    imagePath: '/images/sponsors/jpmorgan-chase-logo.png',
-    altText: 'JPMorgan Chase',
-  },
-  {
-    imagePath: '/images/sponsors/matrix.png',
-    altText: 'Matrix',
-  },
-  {
-    imagePath: '/images/sponsors/paycom.png',
-    altText: 'Paycom',
-  },
-];
+import { sponsorList as Sponsors } from '../data/_data';
+import { useStyles } from '../css/style';
 
 interface SponsorProps {
   imagePath: string;
   altText: string;
 }
 
-const useStyles = createStyles((theme) => ({
-  image: {
-    width: 200,
-  },
-
-  wrapper: {
-    paddingTop: theme.spacing.xl * 4,
-    paddingBottom: theme.spacing.xl * 4,
-  },
-
-  title: {
-    fontFamily: `Greycliff CF, ${theme.fontFamily}`,
-    fontWeight: 800,
-    marginBottom: theme.spacing.md,
-    textAlign: 'center',
-
-    [theme.fn.smallerThan('sm')]: {
-      fontSize: 28,
-      textAlign: 'left',
-    },
-  },
-}));
-
 export function Sponsor({ imagePath, altText } : SponsorProps) {
-  // const classes = useStyles();
   return (
     <div>
       <Center>
@@ -73,14 +27,14 @@ interface FeaturesGridProps {
   data?: SponsorProps[];
 }
 
-export function FeaturesGrid({ data = sponsorList }: FeaturesGridProps) {
+export function FeaturesGrid({ data = Sponsors }: FeaturesGridProps) {
   const { classes, theme } = useStyles();
   const sponsors = data.map((sponsor, index) => <Sponsor {...sponsor} key={index} />);
 
   return (
-    <Container className={classes.wrapper}>
+    <Container className={classes.sponsorWrapper}>
       <Center>
-          <Title className={classes.title}>Special thanks to our{' '}
+          <Title className={classes.sponsorTitle}>Special thanks to our{' '}
             <Text component="span" variant="gradient" gradient={{ from: 'pink', to: 'violet' }} inherit>Partners</Text>
           </Title>
       </Center>
