+++ css/style.js
@@ -1,0 +1,298 @@
+import {createStyles} from "@mantine/core";
+
+export const useStyles = createStyles((theme) => ({
+
+    // ---- Header (Nav Bar) component ----
+
+    header: {
+        display: 'flex',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        height: '100%',
+        border: 'none',
+    },
+
+    links: {
+        [theme.fn.smallerThan('sm')]: {
+            display: 'none',
+        },
+    },
+
+    burger: {
+        [theme.fn.largerThan('sm')]: {
+            display: 'none',
+        },
+        width: '30px',
+    },
+
+    link: {
+        display: 'block',
+        lineHeight: 1,
+        padding: '8px 12px',
+        borderRadius: theme.radius.sm,
+        textDecoration: 'none',
+        color: theme.colorScheme === 'dark' ? theme.colors.dark[0] : theme.colors.gray[7],
+        fontSize: theme.fontSizes.sm,
+        fontWeight: 500,
+
+        '&:hover': {
+            backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
+        },
+    },
+
+    modalLink: {
+        display: 'flex',
+        lineHeight: 1,
+        padding: '8px 12px',
+        borderRadius: theme.radius.sm,
+        textDecoration: 'none',
+        color: theme.colorScheme === 'dark' ? theme.colors.dark[0] : theme.colors.gray[7],
+        fontSize: theme.fontSizes.xl,
+        fontWeight: 700,
+
+        '&:hover': {
+            backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
+        },
+    },
+
+    linkActive: {
+        '&, &:hover': {
+            backgroundColor:
+                theme.colorScheme === 'dark'
+                    ? theme.fn.rgba(theme.colors[theme.primaryColor][9], 0.25)
+                    : theme.colors[theme.primaryColor][0],
+            color: theme.colors[theme.primaryColor][theme.colorScheme === 'dark' ? 3 : 7],
+        },
+    },
+
+    // ---- Hero component ----
+
+    hero: {
+        position: 'relative',
+        // backgroundImage: 'url(https://images.unsplash.com/photo-1519389950473-47ba0277781c?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1770&q=80)',
+        backgroundSize: 'cover',
+        backgroundPosition: 'center',
+    },
+
+    outerContainer: {
+        height: 700,
+        paddingRight: '5rem',
+        display: 'flex',
+        flexDirection: 'row',
+        justifyContent: 'flex-start',
+        alignItems: 'center',
+        [`@media (min-width: ${theme.breakpoints.xl}px)`]: {
+            maxWidth: 1500
+        },
+    },
+
+    container: {
+        height: 700,
+        display: 'flex',
+        flexDirection: 'column',
+        justifyContent: 'flex-end',
+        alignItems: 'flex-start',
+        paddingBottom: theme.spacing.xl * 6,
+        zIndex: 1,
+        position: 'relative',
+
+        [theme.fn.smallerThan('sm')]: {
+            height: 500,
+            paddingBottom: theme.spacing.xl * 3,
+        },
+    },
+
+    image: {
+        width: 550,
+        zIndex: 2,
+        [theme.fn.smallerThan('sm')]: {
+            display: 'none',
+        },
+    },
+
+    heroTitle: {
+        color: theme.black,
+        fontSize: 60,
+        fontWeight: 900,
+        lineHeight: 1.1,
+
+        [theme.fn.smallerThan('sm')]: {
+            fontSize: 40,
+            lineHeight: 1.2,
+        },
+
+        [theme.fn.smallerThan('xs')]: {
+            fontSize: 28,
+            lineHeight: 1.3,
+        },
+    },
+
+    subtitle: {
+        color: theme.black,
+        fontSize: 40,
+        fontWeight: 700,
+        lineHeight: 1.0,
+        marginBottom: '2rem',
+
+        [theme.fn.smallerThan('sm')]: {
+            fontSize: 30,
+            lineHeight: 1.1,
+        },
+
+        [theme.fn.smallerThan('xs')]: {
+            fontSize: 22,
+            lineHeight: 1.1,
+        },
+    },
+
+    boldHeading: {
+        color: theme.black,
+        fontSize: '1.75rem',
+        fontWeight: 800,
+        lineHeight: 1.0,
+        marginBottom: '1rem',
+
+        [theme.fn.smallerThan('sm')]: {
+            fontSize: '1.5rem',
+            lineHeight: 1.0,
+        },
+
+        [theme.fn.smallerThan('xs')]: {
+            fontSize: '1.25rem',
+            lineHeight: 1.0,
+        },
+    },
+
+    heroDescription: {
+        color: theme.black,
+        maxWidth: 600,
+
+        [theme.fn.smallerThan('sm')]: {
+            maxWidth: '100%',
+            fontSize: theme.fontSizes.sm,
+        },
+    },
+
+    control: {
+        marginTop: theme.spacing.xl * 1.5,
+
+        [theme.fn.smallerThan('sm')]: {
+            width: '100%',
+        },
+    },
+
+    // ---- Sponsors component ----
+
+    sponsorWrapper: {
+        paddingTop: theme.spacing.xl * 4,
+        paddingBottom: theme.spacing.xl * 4,
+    },
+
+    sponsorTitle: {
+        fontFamily: `Greycliff CF, ${theme.fontFamily}`,
+        fontWeight: 800,
+        marginBottom: theme.spacing.md,
+        textAlign: 'center',
+
+        [theme.fn.smallerThan('sm')]: {
+            fontSize: 28,
+            textAlign: 'left',
+        },
+    },
+
+    // ---- Footer component ----
+
+    footer: {
+        marginTop: 120,
+        paddingTop: theme.spacing.xl * 2,
+        paddingBottom: theme.spacing.xl * 2,
+        backgroundColor: theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[0],
+        borderTop: `1px solid ${
+            theme.colorScheme === 'dark' ? theme.colors.dark[5] : theme.colors.gray[2]
+        }`,
+    },
+
+    logo: {
+        maxWidth: 200,
+
+        [theme.fn.smallerThan('sm')]: {
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+        },
+    },
+
+    description: {
+        marginTop: 5,
+
+        [theme.fn.smallerThan('sm')]: {
+            marginTop: theme.spacing.xs,
+            textAlign: 'center',
+        },
+    },
+
+    inner: {
+        display: 'flex',
+        justifyContent: 'space-between',
+
+        [theme.fn.smallerThan('sm')]: {
+            flexDirection: 'column',
+            alignItems: 'center',
+        },
+    },
+
+    groups: {
+        display: 'flex',
+        flexWrap: 'wrap',
+
+        [theme.fn.smallerThan('sm')]: {
+            display: 'none',
+        },
+    },
+
+    wrapper: {
+        width: 160,
+    },
+
+    footerLink: {
+        display: 'block',
+        color: theme.colorScheme === 'dark' ? theme.colors.dark[1] : theme.colors.gray[6],
+        fontSize: theme.fontSizes.sm,
+        paddingTop: 3,
+        paddingBottom: 3,
+
+        '&:hover': {
+            textDecoration: 'underline',
+        },
+    },
+
+    title: {
+        fontSize: theme.fontSizes.lg,
+        fontWeight: 700,
+        fontFamily: `Greycliff CF, ${theme.fontFamily}`,
+        marginBottom: theme.spacing.xs / 2,
+        color: theme.colorScheme === 'dark' ? theme.white : theme.black,
+    },
+
+    afterFooter: {
+        display: 'flex',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        marginTop: theme.spacing.xl,
+        paddingTop: theme.spacing.xl,
+        paddingBottom: theme.spacing.xl,
+        borderTop: `1px solid ${
+            theme.colorScheme === 'dark' ? theme.colors.dark[4] : theme.colors.gray[2]
+        }`,
+
+        [theme.fn.smallerThan('sm')]: {
+            flexDirection: 'column',
+        },
+    },
+
+    social: {
+        [theme.fn.smallerThan('sm')]: {
+            marginTop: theme.spacing.xs,
+        },
+    },
+}));
