+++ data/_data.js
@@ -1,0 +1,58 @@
+// Header navigation
+export const navLinks = [
+        {
+            "link": "/",
+            "label": "Home"
+        },
+        {
+            "link": "/about",
+            "label": "About"
+        },
+        {
+            "link": "/tracks",
+            "label": "Tracks"
+        },
+        {
+            "link": "/faq",
+            "label": "F.A.Q."
+        },
+        {
+            "link": "/team",
+            "label": 'Team',
+        },
+        {
+            "link": "/sponsors",
+            "label": 'Sponsors',
+        },
+        {
+            "link": "/guide",
+            "label": 'Guide',
+        },
+];
+
+// Sponsors
+export const sponsorList = [
+    {
+        imagePath: '/images/sponsors/google-cloud.png',
+        altText: 'Google Cloud',
+    },
+    {
+        imagePath: '/images/sponsors/H-E-B.png',
+        altText: 'H-E-B',
+    },
+    {
+        imagePath: '/images/sponsors/jpmorgan-chase-logo.png',
+        altText: 'JPMorgan Chase',
+    },
+    {
+        imagePath: '/images/sponsors/matrix.png',
+        altText: 'Matrix',
+    },
+    {
+        imagePath: '/images/sponsors/paycom.png',
+        altText: 'Paycom',
+    },
+];
+
+// Footer data
+export const footerLinks = { };
