--- data/_links.js
@@ -1,70 +1,0 @@
-// Header navigation
-import {IconCookie, IconGauge, IconLock, IconMessage2, IconUser} from "@tabler/icons";
-
-export const navLinks = [
-        {
-            "link": "/",
-            "label": "Home"
-        },
-        {
-            "link": "/about",
-            "label": "About"
-        },
-        {
-            "link": "/tracks",
-            "label": "Tracks"
-        },
-        {
-            "link": "/faq",
-            "label": "F.A.Q."
-        },
-        {
-            "link": "/team",
-            "label": 'Team',
-        },
-        {
-            "link": "/sponsors",
-            "label": 'Sponsors',
-        },
-        {
-            "link": "/guide",
-            "label": 'Guide',
-        },
-];
-
-// Sponsors
-export const sponsors = [
-    // {
-    //     image: "/images/sponsors/google-cloud.png",
-    //     title: 'Google Cloud',
-    //     description:
-    //         'This dust is actually a powerful poison that will even make a pro wrestler sick, Regice cloaks itself with frigid air of -328 degrees Fahrenheit',
-    // },
-    {
-        icon: IconUser,
-        title: 'Privacy focused',
-        description:
-            'People say it can run at the same speed as lightning striking, Its icy body is so cold, it will not melt even if it is immersed in magma',
-    },
-    {
-        icon: IconCookie,
-        title: 'No third parties',
-        description:
-            'They’re popular, but they’re rare. Trainers who show them off recklessly may be targeted by thieves',
-    },
-    {
-        icon: IconLock,
-        title: 'Secure by default',
-        description:
-            'Although it still can’t fly, its jumping power is outstanding, in Alola the mushrooms on Paras don’t grow up quite right',
-    },
-    {
-        icon: IconMessage2,
-        title: '24/7 Support',
-        description:
-            'Rapidash usually can be seen casually cantering in the fields and plains, Skitty is known to chase around after its own tail',
-    },
-];
-
-// Footer data
-export const footerLinks = { };
